requirements.txt
tensorflow-gpu==1.11
keras
sklearn
matplotlib
numpy
opencv-python